requirements.txt
pip>=0.8.2
South
Django>=1.5
django-extensions
python-dateutil
celery
requests
lxml  # requires libxslt-dev and libxml2-dev system packages
djangorestframework

# optional/suggested
#psycopg2

# install these if you intend to develop
mox
nose
django-nose
fabric
